minor: fixed a compiler warning in H5DatasetHandler_free and H5FileIOHandler_free
reason: freeing a const pointer makes errors as free needs non const pointers. 'name' and 'filename' are const variables to signal they won't change its value after being assigned in the init function. Hence I cast away the const in the free call as it is the only function other than init which changes the contents of the variables.